requirements.txt
fastapi==0.115.12
PyYAML==6.0.2
pymilvus==2.5.10
sentence-transformers==4.1.0
confluent-kafka==2.10.0
streamlit==1.45.1
requests==2.32.3
uvicorn==0.34.3
sentence-transformers==4.1.0
elasticsearch==8.9.0
transformers==4.52.4
dateparser==1.1.8
python-dateutil==2.9.0.post0
psutil==7.0.0
requests==2.32.3
python-dotenv==1.1.0
pydantic==2.8.2
typing-extensions==4.11.0
diskcache==5.6.3
pandas==2.2.0
plotly==5.18.0
locust==2.20.0
